Переписал store на модули
